+++ tests/Helpers/NameHelperTest.php
@@ -8,7 +8,8 @@
  * file that was distributed with this source code.
  */
 
-namespace Test\Helpers;
+
+namespace Tests\Helpers;
 
 use PayBreak\Foundation\Helpers\NameHelper;
 
@@ -16,18 +17,26 @@ use PayBreak\Foundation\Helpers\NameHelper;
  * Name Helper Test
  *
  * @author WN
- * @package Test\Helpers
+ * @package Tests\Helpers
  */
 class NameHelperTest extends \PHPUnit_Framework_TestCase
 {
     public function testCamelToSnake()
     {
         $this->assertSame('xx_ya', NameHelper::camelToSnake('XxYa'));
+        $this->assertSame('xxx_yyy_www', NameHelper::camelToSnake('XxxYyyWww'));
     }
 
     public function testSnakeToCamel()
     {
         $this->assertSame('XxYa', NameHelper::snakeToCamel('xx_ya'));
         $this->assertSame('xxYa', NameHelper::snakeToCamel('xx_ya', true));
+        $this->assertSame('XxxYyyWww', NameHelper::snakeToCamel('xxx_yyy_www'));
+        $this->assertSame('XxxYyyWww', NameHelper::snakeToCamel('xxx_yYy_www'));
+    }
+
+    public function testSnakeToCamelLow()
+    {
+        $this->assertSame('xxxYyyWww', NameHelper::snakeToCamel('xxx_yyy_www', true));
     }
 }