+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Features
+- Added `AbstractEntity` support for `class[]`
+- Added `NameHelper`
+- Improved `trigger_error` and `EntityPropertyTrait`
+- Doc bloc generator for `AbstractEntity`
+
 # v0.8.0
 2016-01-20